+++ src/containers/Generation/GenerationForm/GenerationContributionForm.js
@@ -42,7 +42,7 @@ const useStyles = makeStyles((theme) => ({
 }))
 
 const GenerationContributionForm = (props) => {
-  const { values, handleBlur, errors, touched, setFieldValue } = props
+  const { values, handleBlur, errors, touched, setFieldValue, title } = props
 
   const { t } = useTranslation()
   const classes = useStyles()
@@ -116,14 +116,19 @@ const GenerationContributionForm = (props) => {
 
   return (
     <>
-      <StepHeader title={t('GENERATION_FORM_TITLE')} />
+      <Typography component="h1" variant="h3">
+        {title}
+      </Typography>
       <Typography variant="h6">
         {t('GENERATION_FORM_ANNUAL_USE_TITLE')}
       </Typography>
       <Box pt={1}>
-        <Typography variant="body1">
-          {t('GENERATION_FORM_ANNUAL_USE_INPUT_TITLE')}
-        </Typography>
+        <Typography variant="body1"
+          dangerouslySetInnerHTML={{
+            __html: t('GENERATION_FORM_ANNUAL_USE_INPUT_TITLE', {
+              url: t('GENERATION_FORM_ANNUAL_CONSUMPTION_URL')
+            })
+          }} />
       </Box>
       <Box pt={1} id="box_annual_use">
         <TextField
@@ -140,8 +145,7 @@ const GenerationContributionForm = (props) => {
           onBlur={handleBlur}
           error={errors?.annual_use && touched?.annual_use}
           helperText={
-            (touched?.annual_use && errors?.annual_use) ||
-            t('GENERATION_FORM_ANNUAL_USE_INPUT_HELP_TEXT')
+            (touched?.annual_use && errors?.annual_use) || ''
           }
         />
       </Box>